Resolve the issue in the chained operation
Before (1+1+2-1) gives 1 instead of doing each operation.
I declare a operatorFirstClick to check wether it's
the first time an operator is being clicked. If true,
the the firstNumber is saved and the operator too and we wait
for the second number. If false, it take the secondNumber
on the display and call the operate function, and then
save the result in the first number for the next operation, and wait for
the secondNumber for the operation. it also use the
new operator.